refactor: change users methods from package
