+++ grails-app/controllers/com/mini/asaas/customer/CustomerController.groovy
@@ -5,10 +5,6 @@ import com.mini.asaas.dto.customer.CustomerDTO
 import grails.compiler.GrailsCompileStatic
 import grails.validation.ValidationException
 import grails.plugin.springsecurity.annotation.Secured
-import com.mini.asaas.user.User
-import com.mini.asaas.user.Role
-import com.mini.asaas.repositories.UserRepository
-import com.mini.asaas.repositories.RoleRepository
 import com.mini.asaas.user.UserService
 import com.mini.asaas.dto.user.UserDTO
 import com.mini.asaas.email.EmailService
@@ -22,8 +18,6 @@ class CustomerController {
     AddressService addressService
 
     UserService userService
-
-    EmailService emailService
 
     def index() {}
 
@@ -70,43 +64,6 @@ class CustomerController {
             log.error("CustomerController.update >> Não foi possível atualizar o Customer", exception)
             params.errorMessage = "Não foi possível editar o cliente"
             redirect(action: "edit", params: params)
-        }
-    }
-
-    def users() {
-        try {
-            Customer customer = userService.getCustomerByUser()
-            List<User> userList = UserRepository.listByCustomer(customer.id)
-            List<Role> roleList = RoleRepository.listAll()
-            return [userList: userList, roleList: roleList]
-        } catch (Exception exception) {
-            log.error("CustomerController.users >> Não foi possível listar os Users", exception)
-            params.errorMessage = "Não foi possível listar os usuários"
-            redirect(view: "edit", params: params)
-        }
-    }
-
-    def addUser() {
-        try {
-            Random randomPassword = new Random()
-            int maxLenght = 999999
-            params.password = randomPassword.nextInt(maxLenght)
-
-            Customer customer = userService.getCustomerByUser()
-            params.customer = customer
-
-            UserDTO userDTO = new UserDTO(params)
-
-            Role role = Role.findByAuthority(params.role)
-            User user = userService.save(userDTO, role)
-
-            emailService.notifyOnNewUser(user, params.password.toString())
-
-            redirect(action: 'users')
-        } catch (Exception exception) {
-            log.error("RegisterController.register >> Não foi possível registrar o usuário", exception)
-            params.errorMessage = "Não foi possível registrar o usuário"
-            redirect(action: 'edit', params: params)
         }
     }
 }
