--- grails-app/views/customer/users.gsp
@@ -1,25 +1,0 @@
-<!doctype html>
-<html>
-<head>
-</head>
-<body>
-    <div>
-        <h2>Usuários</h2>
-        <g:each var="user" in="${userList}">
-            <span>Email: ${user.username}</span><br>
-        </g:each> 
-            
-        <g:form action="addUser" method="post">
-            <h3>Adicionar Usuário</h3>
-
-            <label for="email">Email:</label>
-            <g:textField name="email" required="true"/><br>
-            
-            <label for="role">Função:</label>
-            <g:select name="role" from="${roleList.authority}" valueMessagePrefix="Role" required="true"/><br><br>
-    
-            <g:submitButton name="add" value="Adicionar" />
-        </g:form>        
-    </div>
-</body>
-</html>
